Add files via upload
This is the config.yml file that was developed for the linear motors in the TOI lab. A config.yml file is needed to run the motor in the GUI. However, one can develop their own config.yml file through the official pytic Github page. This config.yml file was not developed by the author of this repository.